Move keyvault common headers to the right directory
