+++ sdk/keyvault/azure-security-keyvault-common/inc/azure/keyvault/common/internal/keyvault_exception.hpp
@@ -15,7 +15,7 @@
 #include <stdexcept>
 #include <string>
 
-namespace Azure { namespace Security { namespace KeyVault { namespace _detail {
+namespace Azure { namespace Security { namespace KeyVault { namespace _internal {
   /**
    * @brief Container for static methods to parse Key Vault payloads to Azure Core Exception.
    *
