+++ sdk/keyvault/azure-security-keyvault-common/src/keyvault_exception.cpp
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/keyvault/common/keyvault_exception.hpp"
-#include "azure/keyvault/common/keyvault_constants.hpp"
+#include "azure/keyvault/common/internal/keyvault_exception.hpp"
+#include "private/keyvault_constants.hpp"
 
 #include <azure/core/http/policies/policy.hpp>
 
@@ -12,7 +12,7 @@
 using namespace Azure::Security::KeyVault;
 using namespace Azure::Core::Http::_internal;
 
-Azure::Core::RequestFailedException _detail::KeyVaultException::CreateException(
+Azure::Core::RequestFailedException _internal::KeyVaultException::CreateException(
     std::unique_ptr<Azure::Core::Http::RawResponse> rawResponse)
 {
   std::vector<uint8_t> bodyBuffer = std::move(rawResponse->GetBody());
