+++ sdk/keyvault/azure-security-keyvault-common/src/private/keyvault_constants.hpp
