+++ sdk/keyvault/azure-security-keyvault-keys/src/recover_deleted_key_operation.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <azure/keyvault/common/keyvault_exception.hpp>
+#include <azure/keyvault/common/internal/keyvault_exception.hpp>
 
 #include "azure/keyvault/keys/details/key_constants.hpp"
 #include "azure/keyvault/keys/details/key_serializers.hpp"
@@ -40,7 +40,7 @@ Azure::Security::KeyVault::Keys::RecoverDeletedKeyOperation::PollInternal(
         break;
       }
       default:
-        throw Azure::Security::KeyVault::_detail::KeyVaultException::CreateException(
+        throw Azure::Security::KeyVault::_internal::KeyVaultException::CreateException(
             std::move(rawResponse));
     }
     if (m_status == Azure::Core::OperationStatus::Succeeded)
